*: move all non-Godep code into src/ subdirectory
Introduces a top-level config->src/config symlink to preserve the clean
"github.com/coreos/ignition/config" import path; config is intended to have
external users depend upon it.
