+++ src/exec/stages/prepivot/prepivot.go
@@ -18,9 +18,9 @@ import (
 	"fmt"
 
 	"github.com/coreos/ignition/config"
-	"github.com/coreos/ignition/exec/stages"
-	"github.com/coreos/ignition/exec/util"
-	"github.com/coreos/ignition/log"
+	"github.com/coreos/ignition/src/exec/stages"
+	"github.com/coreos/ignition/src/exec/util"
+	"github.com/coreos/ignition/src/log"
 )
 
 const (
