+++ src/exec/stages/stages.go
@@ -16,8 +16,8 @@ package stages
 
 import (
 	"github.com/coreos/ignition/config"
-	"github.com/coreos/ignition/log"
-	"github.com/coreos/ignition/registry"
+	"github.com/coreos/ignition/src/log"
+	"github.com/coreos/ignition/src/registry"
 )
 
 type Stage interface {
