+++ src/log/stdout.go
