+++ src/providers/cmdline/cmdline.go
@@ -22,9 +22,9 @@ import (
 	"time"
 
 	"github.com/coreos/ignition/config"
-	"github.com/coreos/ignition/log"
-	"github.com/coreos/ignition/providers"
-	"github.com/coreos/ignition/providers/util"
+	"github.com/coreos/ignition/src/log"
+	"github.com/coreos/ignition/src/providers"
+	"github.com/coreos/ignition/src/providers/util"
 )
 
 const (
