+++ src/providers/providers.go
@@ -18,8 +18,8 @@ import (
 	"time"
 
 	"github.com/coreos/ignition/config"
-	"github.com/coreos/ignition/log"
-	"github.com/coreos/ignition/registry"
+	"github.com/coreos/ignition/src/log"
+	"github.com/coreos/ignition/src/registry"
 )
 
 type Provider interface {
