+++ src/registry/registry.go
