+++ src/registry/registry_test.go
